Factor out a common `SettingsSubHeader` component
